% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procs.R
\name{fdtest}
\alias{fdtest}
\title{Randomization test}
\usage{
fdtest(fc, functional = "expectile", alpha = 0.5, np = 1000,
  test_functional = "sqpos")
}
\arguments{
\item{fc}{matrix with forecasts (first two cols) and realization (third col)}

\item{functional}{either "expectile" or "quantile"}

\item{alpha}{level of the expectile or quantile}

\item{np}{nr of randomization draws}

\item{test_functional}{test statistic to be used, either "sqpos" or "abspos"}
}
\description{
Randomization test for forecast dominance
}
\details{
\emph{Note:} The test's H0 is that forecast 1 weakly dominates forecast 2. Hence, low p-values indicate that forecast 1 is \emph{not} dominant. This can mean that either i) forecast 2 is dominant or that ii) none of the forecasts dominates the other.
}
