% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procs.R
\name{sim_dgp}
\alias{sim_dgp}
\alias{sim_dgp_est}
\alias{sim_dgp_quantiles}
\alias{sim_dgp_quantiles,}
\alias{sim_dgp_est}
\title{Functions to simulate example data}
\usage{
sim_dgp(n, g = 0.8, h = 1, sd_2 = 1)

sim_dgp_est(n, g = 0.8, sd_2 = 1, presample = 50)

sim_dgp_quantiles(n, a = 0.05, tau1, tau2)
}
\arguments{
\item{n}{nr of observations}

\item{g}{persistence parameter}

\item{h}{forecast horizon}

\item{sd_2}{variance parameter, determines relative performance}

\item{presample, }{number of presample observations}

\item{a}{quantile level}

\item{tau1}{precision parameter for forecaster 1}

\item{tau2}{precision parameter for forecaster 2}
}
\value{
matrix of dimension (n x 3), containing forecasts (columns 1, 2) and realizations (column 3) from example process
}
\description{
Functions to simulate example data
}
